TASK6

+++ app/Http/Controllers/StudentContoller.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Student;
+use Hamcrest\Type\IsNumeric;
 use Illuminate\Http\Request;
+
+use function PHPUnit\Framework\isNull;
 
 class StudentController extends Controller
 {
@@ -12,11 +15,35 @@ class StudentController extends Controller
     {
         $students = Student::all();
 
-        $data = [
-            'message' => 'Get all students',
-            'data' => $students,
-        ];
+        if ($students->count() == 0) {
+            $data = [
+                'message' => 'Data student is empty'
+            ];
+            return response()->json($data, 404);
+        } else {
+            $data = [
+                'message' => 'Get all students',
+                'data' => $students,
+            ];
+            return response()->json($data, 200);
+        }
+    }
 
+    public function show($id)
+    {
+        $data = Student::find($id);
+
+        if (!is_numeric($id)) {
+            $errorMessage = [
+                'message' => "id you entered '$id' is not a number"
+            ];
+            return response()->json($errorMessage, 404);
+        } else if ($data == null) {
+            $errorMessage = [
+                'message' => "data with id $id doesn't exist"
+            ];
+            return response()->json($errorMessage, 404);
+        }
         return response()->json($data, 200);
     }
 
@@ -29,20 +56,35 @@ class StudentController extends Controller
         $email = $request->email;
         $jurusan = $request->jurusan;
 
-        // Insert data to database -> students
-        $student = Student::create([
-            "nama" => $nama,
-            "nim" => $nim,
-            "email" => $email,
-            "jurusan" => $jurusan
-        ]);
+        if ($nama && $nim && $email && $jurusan) {
+            // Insert data to database -> students
+            $student = Student::create([
+                "nama" => $nama,
+                "nim" => $nim,
+                "email" => $email,
+                "jurusan" => $jurusan
+            ]);
 
-        $data = [
-            "message" => "Student data successfully created",
-            "data" => $student
-        ];
+            $data = [
+                "message" => "Student data successfully created",
+                "data" => $student
+            ];
 
-        return response()->json($data, 201);
+            return response()->json($data, 201);
+        } else {
+            $errorMessages = [];
+
+            $nama ?? array_push($errorMessages, "Nama has not been filled");
+            $nim ?? array_push($errorMessages, "NIM has not been filled");
+            $email ?? array_push($errorMessages, "Email has not been filled");
+            $jurusan ?? array_push($errorMessages, "Jurusan has not been filled");
+            
+            $data = [
+                "Add data student failed" => $errorMessages
+            ];
+
+            return response()->json($data, 404);
+        }
     }
 
     // Edit student data
@@ -56,32 +98,45 @@ class StudentController extends Controller
 
         // Update student data
         $student = Student::find($id);
-        $student->update([
-            'nama' => $nama,
-            'nim' => $nim,
-            'email' => $email,
-            'jurusan' => $jurusan,
-        ]);
 
-        $data = [
-            "message" => "Student with id $id has succesfully updated",
-            "data" => $student
-        ];
+        if ($student) {
+            $student->update([
+                'nama' => ($nama != null) ? $nama : $student->nama,
+                'nim' => ($nim != null) ? $nim : $student->nim,
+                'email' => ($email != null) ? $email : $student->email,
+                'jurusan' => ($jurusan != null) ? $jurusan : $student->jurusan,
+            ]);
 
-        return response()->json($data, 200);
+            $data = [
+                "message" => "Student with id $id has succesfully updated",
+                "data" => $student
+            ];
+            return response()->json($data, 200);
+        } else {
+            $errorMessage = [
+                "message" => "update data failed, data with id $id doesn't exist"
+            ];
+            return response()->json($errorMessage, 404);
+        }
     }
 
     // Delete student data
     public function delete($id)
     {
         $student = Student::find($id);
-        $student->delete();
 
-        $data = [
-            "message" => "Student with id $id has succesfully deleted",
-            "data" => $student
-        ];
-
-        return response()->json($data, 200);
+        if ($student) {
+            $student->delete();
+            $data = [
+                "message" => "Student with id $id has succesfully deleted",
+                "data" => $student
+            ];
+            return response()->json($data, 200);
+        } else {
+            $data = [
+                "message" => "delete data failed, data with id $id doesn't exist"
+            ];
+            return response()->json($data, 404);
+        }
     }
 }
